✨ [Feat] refactor pending quotes API: remove old file and implement new structure
